Move the 'omnibus' test into a separate file
